Traffic_Sign_Detection-Gradio&Streamlit_Apps/requirements.txt
pandas
numpy
requests
pillow
scikit-learn
tensorflow
keras
